* Rename the package of the new migration files * Add comments for UserInfoBackup.java * Make UserService listen to UserSessionLoginEvent and   record user's details in here
