+++ Source/Plugins/Core/com.equella.core/resources/lang/i18n-resource-centre.properties
@@ -76,8 +76,8 @@
 /com.tle.core.entity.services.migration.v64.AddInstitutionFileWarningColumn.title=Add quota warning column to institution
 /com.tle.core.entity.services.migration.v64.displaytemplatetitle.title=Add new title to display template entities
 /com.tle.core.entity.services.migration.v64.thumbnail.columnsize.title=Increasing column size for item thumbnails
-/com.tle.core.entity.services.migration.v65.unknownuser.lastowner=Add last owner to item
-/com.tle.core.entity.services.migration.v65.unknownuser.alluser=Create a new table for all users
+/com.tle.core.entity.services.migration.v20192.unknownuser.lastowner=Add last owner to item
+/com.tle.core.entity.services.migration.v20192.unknownuser.alluser=Create a new table for all users
 /com.tle.core.entity.services.query.contains={0} is {1}
 /com.tle.core.entity.services.query.date.after={0} after {1}
 /com.tle.core.entity.services.query.date.before={0} before {1}
